feat(dockerfile): back to root
Permissions are a pain. Having a nonroot user is more secure, but having to change permissions on everything also changes them on the host filesystem if you map the volume which introduces all kinds of headaches. 

As this is not a long running container but always a temporary one that only runs when I work this is an acceptable tradeoff.